don't ovewrite Dockerfile on restart

+++ src/install.sh
@@ -5,7 +5,7 @@ clear
 # copy install files
 [ ! -f ~/localhost/hosts.conf ] && cp  ~/localhost/repo/src/hosts.conf ~/localhost/hosts.conf
 [ ! -d ~/localhost/www ] && cp -R  ~/localhost/repo/www ~/localhost
-cp ~/localhost/repo/src/Dockerfile ~/localhost
+[ ! -f ~/localhost/Dockerfile ] && cp ~/localhost/repo/src/Dockerfile ~/localhost
 
 # remove old image
 docker kill server_container &>/dev/null
